requirements.txt
Django==3.2.4
djangorestframework==3.12.4
numpy==1.21.0
psycopg2-binary==2.9.1
python-decouple==3.4